Add files via upload
This is web portal in which users are registered and enter their and own family medical history to track their fitness issues and medicinal drugs in exclusive durations. In this project we solve the major problems of every patients because we manage the whole record in his life. In every moment when he want to check up he will be able to see the previous record easily. The major things is that doctor can easily understand the patient disease with the help of this portal